LOG-4812: Refactor metrics templates and helper functions
