+++ internal/generator/vector/normalize/throttle.go
@@ -1,11 +1,29 @@
 package normalize
 
+import (
+	"github.com/openshift/cluster-logging-operator/internal/generator/framework"
+	"github.com/openshift/cluster-logging-operator/internal/generator/vector/helpers"
+)
+
 type Throttle struct {
 	ComponentID string
 	Desc        string
 	Inputs      string
 	Threshold   int64
 	KeyField    string
+}
+
+func NewThrottle(id string, inputs []string, threshhold int64, throttleKey string) []framework.Element {
+	el := []framework.Element{}
+
+	el = append(el, Throttle{
+		ComponentID: id,
+		Inputs:      helpers.MakeInputs(inputs...),
+		Threshold:   threshhold,
+		KeyField:    throttleKey,
+	})
+
+	return el
 }
 
 func (t Throttle) Name() string {
