+++ internal/generator/vector/source/metrics.go
@@ -1,10 +1,7 @@
 package source
 
 const (
-	InternalMetricsSourceName        = "internal_metrics"
-	PrometheusOutputSinkName         = "prometheus_output"
-	PrometheusExporterListenPort     = `24231`
-	AddNodenameToMetricTransformName = "add_nodename_to_metric"
+	InternalMetricsSourceName = "internal_metrics"
 )
 
 type InternalMetrics struct {
@@ -25,53 +22,4 @@ func (i InternalMetrics) Template() string {
 type = "internal_metrics"
 {{end}}
 `
-}
-
-type PrometheusExporter struct {
-	ID            string
-	Inputs        string
-	Address       string
-	TlsMinVersion string
-	CipherSuites  string
-}
-
-func (p PrometheusExporter) Name() string {
-	return "PrometheusExporterTemplate"
-}
-
-func (p PrometheusExporter) Template() string {
-	return `{{define "` + p.Name() + `" -}}
-[sinks.{{.ID}}]
-type = "prometheus_exporter"
-inputs = {{.Inputs}}
-address = "{{.Address}}"
-default_namespace = "collector"
-
-[sinks.{{.ID}}.tls]
-enabled = true
-key_file = "/etc/collector/metrics/tls.key"
-crt_file = "/etc/collector/metrics/tls.crt"
-min_tls_version = "{{.TlsMinVersion}}"
-ciphersuites = "{{.CipherSuites}}"
-{{end}}`
-}
-
-type AddNodenameToMetric struct {
-	ID     string
-	Inputs string
-}
-
-func (a AddNodenameToMetric) Name() string {
-	return AddNodenameToMetricTransformName
-}
-
-func (a AddNodenameToMetric) Template() string {
-	return `{{define "` + a.Name() + `" -}}
-[transforms.{{.ID}}]
-type = "remap"
-inputs = {{.Inputs}}
-source = '''
-.tags.hostname = get_env_var!("VECTOR_SELF_NODE_NAME")
-'''
-{{end}}`
 }
